Tests: Move contents of tests/conftest.py
